Day 07 Clean up - Day 09 solution wip
